+++ packages/admin-ui/ui/src/domain/gift-cards/details/edit-gift-card-modal.tsx
@@ -53,7 +53,7 @@ const EditGiftCardModal = ({
         onSuccess: () => {
           notification(
             "Updated Gift card",
-            "Gift card was succesfully updated",
+            "Gift card was successfully updated",
             "success"
           )