+++ www/api-reference/components/Navbar/MenuButton/index.tsx
@@ -5,20 +5,22 @@ import { useSidebar } from "@/providers/sidebar"
 import IconSidebar from "../../Icons/Sidebar"
 import clsx from "clsx"
 import IconXMark from "../../Icons/XMark"
+import { usePageLoading } from "../../../providers/page-loading"
 
 type NavbarMenuButtonProps = {
   buttonProps?: NavbarIconButtonProps
 }
 
 const NavbarMenuButton = ({ buttonProps }: NavbarMenuButtonProps) => {
-  const { items, setMobileSidebarOpen, mobileSidebarOpen } = useSidebar()
+  const { setMobileSidebarOpen, mobileSidebarOpen } = useSidebar()
+  const { isLoading } = usePageLoading()
 
   return (
     <NavbarIconButton
       {...buttonProps}
       className={clsx("mr-1 lg:!hidden", buttonProps?.className)}
       onClick={() => {
-        if (items.top.length !== 0 && items.bottom.length !== 0) {
+        if (!isLoading) {
           setMobileSidebarOpen((prevValue) => !prevValue)
         }
       }}