+++ www/docs/sidebars.js
@@ -57,6 +57,16 @@ module.exports = {
           id: "recipes/subscriptions",
           label: "Subscriptions",
         },
+        {
+          type: "doc",
+          id: "recipes/b2b",
+          label: "B2B / Wholesale",
+        },
+        {
+          type: "doc",
+          id: "recipes/multi-region",
+          label: "Multi-Region Store",
+        },
       ],
     },
     {
@@ -220,6 +230,15 @@ module.exports = {
       className: "homepage-sidebar-item",
     },
     {
+      type: "doc",
+      id: "beta",
+      label: "Beta Features",
+      customProps: {
+        sidebar_icon: "adjustments",
+      },
+      className: "homepage-sidebar-item",
+    },
+    {
       type: "category",
       label: "Deploy",
       customProps: {
@@ -357,6 +376,15 @@ module.exports = {
       label: "Contribution Guidelines",
       customProps: {
         sidebar_icon: "document-text",
+      },
+      className: "homepage-sidebar-item",
+    },
+    {
+      type: "doc",
+      id: "glossary",
+      label: "Glossary",
+      customProps: {
+        sidebar_icon: "book-open",
       },
       className: "homepage-sidebar-item",
     },
@@ -1129,6 +1157,11 @@ module.exports = {
           type: "doc",
           id: "modules/users/backend/send-invite",
           label: "Backend: Send Invite",
+        },
+        {
+          type: "doc",
+          id: "modules/users/backend/rbac",
+          label: "Backend: Implement RBAC",
         },
         {
           type: "doc",
@@ -2253,6 +2286,30 @@ module.exports = {
               iconName: "bolt-solid",
               description:
                 "Learn how to use local file storage in your Medusa backend",
+            },
+          },
+        ],
+      },
+      {
+        type: "category",
+        label: "Other",
+        collapsible: false,
+        link: {
+          type: "doc",
+          id: "plugins/other/index",
+        },
+        customProps: {
+          sidebar_is_group_headline: true,
+        },
+        items: [
+          {
+            type: "doc",
+            id: "plugins/other/ip-lookup",
+            label: "IP Lookup",
+            customProps: {
+              iconName: "bolt-solid",
+              description:
+                "Learn how to integrate ipstack to access the user's region.",
             },
           },
         ],