+++ www/docs/src/components/TextArea/index.tsx
@@ -14,7 +14,7 @@ const TextArea: React.FC<TextAreaProps> = (props) => {
       {...props}
       className={clsx(
         "bg-medusa-bg-field dark:bg-medusa-bg-field-dark shadow-button-secondary dark:shadow-button-secondary-dark",
-        "rounded border-medusa-border-loud-muted dark:border-medusa-border-loud-muted-dark border border-solid",
+        "rounded border-medusa-border-base dark:border-medusa-border-base-dark border border-solid",
         "px-0.75 pt-0.4 pb-[9px]",
         "hover:bg-medusa-bg-field-hover dark:hover:bg-medusa-bg-field-hover-dark",
         "focus:border-medusa-border-interactive dark:focus:border-medusa-border-interactive-dark",