+++ www/docs/src/types/index.d.ts
@@ -131,6 +131,12 @@ declare module "@medusajs/docs" {
 
   export declare type NavbarAction = NavbarActionLink | NavbarActionButton
 
+  export declare type OptionType = {
+    value: string
+    label: string
+    isAllOption?: boolean
+  }
+
   export declare type ThemeConfig = {
     reportCodeLinkPrefix?: string
     footerFeedback: {
@@ -154,6 +160,20 @@ declare module "@medusajs/docs" {
       magicComments: MagicCommentConfig[]
     }
     mobileLogo: NavbarLogo
+    algoliaConfig?: {
+      apiKey: string
+      indexNames: {
+        docs: string
+        api: string
+      }
+      appId: string
+      filters: OptionType[]
+      defaultFilters: string[]
+      defaultFiltersByPath: {
+        path: string
+        filters: string[]
+      }[]
+    }
   } & DocusaurusThemeConfig
 
   export declare type MedusaDocusaurusConfig = {